+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - Default `Action#call` param to new context
+- Default failure message parameter to nil
 
 ## [0.2.0] - 2022-09-06